Modularize resolvers.
