+++ src/resolvers/index.ts
@@ -1,3 +1,7 @@
 import Query from './Query';
+import Recipe from './Recipe';
+import { Filters, DietEnum, HealthEnum } from './Filters';
 
-export default { Query };
+export default {
+  Query, Recipe, Filters, DietEnum, HealthEnum,
+};
